move panelReizing into hook
